Major refactoring complete; added threading to user interface; added download speed calculator
