+++ GopherClient/Configuration/GopherClientSettings.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace NetGopherClient.Gopher
+{
+    class ClientSettings
+    {
+        public string textFileViewer { get; set; }
+
+        public bool ChangeLineEncodings { get; set; }
+
+        public bool useAltDowloadForTextFiles { get; set; }
+
+        public List<String> Bookmarks { get; set; }
+    }
+}
