+++ NetGopherClient.sln
@@ -3,12 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetGopherClient", "NetGopherClient\NetGopherClient.csproj", "{F0991151-87EC-43FF-A864-40F495A2BFD3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetGopherClient.Desktop", "NetGopherClient\NetGopherClient.Desktop.csproj", "{F0991151-87EC-43FF-A864-40F495A2BFD3}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{3551A718-E80E-4F42-B517-473DAFB76ADA}"
 	ProjectSection(SolutionItems) = preProject
 		readme.md = readme.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NetGopherClient.Gopher", "GopherClient\NetGopherClient.Gopher.csproj", "{6021B990-54A2-45B6-99EC-7F719FCBC14F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,6 +22,10 @@ Global
 		{F0991151-87EC-43FF-A864-40F495A2BFD3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F0991151-87EC-43FF-A864-40F495A2BFD3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F0991151-87EC-43FF-A864-40F495A2BFD3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6021B990-54A2-45B6-99EC-7F719FCBC14F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6021B990-54A2-45B6-99EC-7F719FCBC14F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6021B990-54A2-45B6-99EC-7F719FCBC14F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6021B990-54A2-45B6-99EC-7F719FCBC14F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
