+++ NetGopherClient/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="NetGopherClient.App"
+﻿<Application x:Class="NetGopherClient.Desktop.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             StartupUri="NetGopherClientWindow.xaml">
+             StartupUri="Windows/NetGopherClientWindow.xaml">
     <Application.Resources>
         <!-- <ResourceDictionary Source="WhistlerBlue.xaml" />
         <ResourceDictionary Source="/PresentationFramework.Aero, Version=3.0.0.0,
