+++ NetGopherClient/Entities/bookmarks.xml
@@ -1,0 +1,3 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+
+<bookmarks></bookmarks>
