+++ NetGopherClient/Windows/FileDownloadWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="NetGopherClient.FileDownloadWindow"
+﻿<Window x:Class="NetGopherClient.Desktop.FileDownloadWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="File Download" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" Width="476" Height="171">
