--- NetGopherClient/bookmarks.xml
@@ -1,3 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-
-<bookmarks></bookmarks>
