--- NetGopherClient/navbuttons.wpf.xaml
@@ -1,6 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:src="clr-namespace:NetGopherClient">
-
-    <!-- This is useless now. -->
-</ResourceDictionary>
